Add GPS and cell metrics sampling for each measurement type (latency, upload, download)

+++ app/src/main/java/com/cellwatch/data/local/model/LatencyDataEntity.kt
@@ -25,13 +25,13 @@ data class LatencyDataEntity(
     @ColumnInfo(index = true)
     var measurementId: String? = null, // UUID
 
-    val rtt: Long? = null,
+    val rtt: Int? = null,
 
-    val jitter: Long? = null,
+    val jitter: Int? = null,
 
-    val sent: Long? = null,
+    val sent: Int? = null,
 
-    val received: Long? = null,
+    val received: Int? = null,
 
     val servers: List<String>? = null,
